rename route for eligibility test for may
