--- site/src/app/v2/test-eligibilite-juin/components/types/types.ts
@@ -1,11 +1,0 @@
-export enum ALLOWANCE {
-  NONE = 'NONE',
-  ARS_AEEH_AAH = 'ARS_AEEH_AAH',
-  CROUS = 'CROUS',
-}
-
-export enum CROUS_AGE_RANGE {
-  LESS_THAN_28 = 'LESS_THAN_28',
-  MORE_THAN_28 = 'MORE_THAN_28',
-  MORE_THAN_29 = 'MORE_THAN_29',
-}
