--- site/src/app/v2/test-eligibilite-juin/page.tsx
@@ -1,21 +1,0 @@
-import AllowanceStep from './components/allowance-step/AllowanceStep';
-import styles from './styles.module.scss';
-import cn from 'classnames';
-
-const EligibilityTest = () => {
-  return (
-    <main className={styles.main}>
-      <h1 className={cn('fr-pt-2w', 'fr-pb-3w', 'fr-px-2w', styles.title)}>
-        Puis-je bénéficier du pass Sport?
-      </h1>
-
-      <div className={cn('fr-mb-8w', 'fr-mx-auto', 'fr-px-2w', styles.background)}>
-        <div className={cn('fr-pt-7w', 'fr-pb-0', 'fr-mx-auto', styles.wrapper)}>
-          <AllowanceStep />
-        </div>
-      </div>
-    </main>
-  );
-};
-
-export default EligibilityTest;
