--- site/src/app/v2/test-eligibilite-juin/styles.module.scss
@@ -1,17 +1,0 @@
-.main {
-  background-color: #ececfe;
-}
-
-.title {
-  text-align: center;
-}
-
-.background {
-  background-color: var(--color-white);
-  max-width: 1200px;
-  min-height: 740px;
-}
-
-.wrapper {
-  max-width: 588px;
-}
