+++ site/src/app/v2/test-eligibilite-mai/styles.module.scss
@@ -1,0 +1,13 @@
+.wrapper {
+  max-width: 588px;
+}
+
+.background {
+  background-color: var(--background-alt-blue-france);
+  max-width: 1200px;
+  min-height: 740px;
+}
+
+.title {
+  text-align: center;
+}
