+++ site/src/app/v2/test-eligibilite/components/qrcode-verdict/QrCOdeVerdict.tsx
