+++ site/src/app/v2/test-eligibilite/components/qrcode-verdict/styles.module.scss
