+++ site/src/app/v2/test-eligibilite/components/types/types.ts
@@ -1,11 +1,11 @@
-export enum AGE_RANGE {
-  BETWEEN_6_19 = 'BETWEEN_6_19',
-  BETWEEN_19_30 = 'BETWEEN_19_30',
-  GREATER_THAN_30 = 'GREATER_THAN_30',
+export enum ALLOWANCE {
+  NONE = 'NONE',
+  ARS_AEEH_AAH = 'ARS_AEEH_AAH',
+  CROUS = 'CROUS',
 }
 
-export enum CHILD_AGE {
-  LESS_THAN_SIX = 'LESS_THAN_SIX',
-  BTW_SIX_AND_THIRTY = 'BTW_SIX_AND_THIRTY',
-  MORE_THAN_THIRTY = 'MORE_THAN_THIRTY',
+export enum CROUS_AGE_RANGE {
+  LESS_THAN_28 = 'LESS_THAN_28',
+  MORE_THAN_28 = 'MORE_THAN_28',
+  MORE_THAN_29 = 'MORE_THAN_29',
 }
