+++ site/src/app/v2/test-eligibilite/page.tsx
@@ -1,6 +1,7 @@
 import { Metadata } from 'next';
-import ForWhoStep from './components/forWhoStep/ForWhoStep';
+import AllowanceStep from './components/allowance-step/AllowanceStep';
 import styles from './styles.module.scss';
+import cn from 'classnames';
 
 export const metadata: Metadata = {
   title: "Test d'éligibilité - pass Sport",
@@ -8,17 +9,17 @@ export const metadata: Metadata = {
 
 const EligibilityTest = () => {
   return (
-    <div>
-      <h1 className={`fr-mt-8w fr-mb-4w fr-px-2w ${styles.title}`}>
+    <main className={styles.main}>
+      <h1 className={cn('fr-pt-2w', 'fr-pb-3w', 'fr-px-2w', styles.title)}>
         Puis-je bénéficier du pass Sport?
       </h1>
 
-      <div className={`fr-mb-8w fr-mx-auto fr-px-2w ${styles.background}`}>
-        <div className={`fr-py-7w fr-mx-auto ${styles.wrapper}`}>
-          <ForWhoStep />
+      <div className={cn('fr-mb-8w', 'fr-mx-auto', 'fr-px-2w', styles.background)}>
+        <div className={cn('fr-pt-7w', 'fr-pb-0', 'fr-mx-auto', styles.wrapper)}>
+          <AllowanceStep />
         </div>
       </div>
-    </div>
+    </main>
   );
 };
