+++ site/src/app/v2/test-eligibilite/styles.module.scss
@@ -1,13 +1,17 @@
-.wrapper {
-  max-width: 588px;
-}
-
-.background {
-  background-color: var(--background-alt-blue-france);
-  max-width: 1200px;
-  min-height: 740px;
+.main {
+  background-color: #ececfe;
 }
 
 .title {
   text-align: center;
+}
+
+.background {
+  background-color: var(--color-white);
+  max-width: 1200px;
+  min-height: 740px;
+}
+
+.wrapper {
+  max-width: 588px;
 }
